New Utility: Its now possible to stop a forEach()-loop
Soo, next try ;) again some utility added and again "return false" is involved… sorry for that..
But this time its a different use of it:

When returning false inside the forEach callback function the loop now stops.
Foreach then returns false else (on success) the forEach function returns true (just for consistency).

Thats it. So generally the function can be used exactly the same as before. But it wasn't possible to stop the loop which can be important.

I also made sure that this is passed along properly in every other use case of the forEach function in basil.js.

PS.: This behavior has also been battle proven in jQuery.