Safety checks to forEach() and typo() + nice little bug fix in items()...
Line 641: pageItems returns a link to the actual collection… this is
updated as soon as e.g. new stuff is drawn to the page and also the new
stuff gets processed... We were stuck in an infinite loop because of
this! allPageItems is returning an array dump of the collection. Keep
this situation in mind for other scenarios.