Added check if given gradient name already exists
1) If a color of the name already existed, an error is thrown. This is because a gradient cannot
be added if it has the same name as any already existing swatch (no matter if it's a gradient
or a color etc.). To not overwrite colors and not run into problems if a name of a default color
is given it makes sense to throw a basil error and let the user know to choose a different name
than the color

2) If a gradient of the same name already existed, it is assumed that the old gradient should be
replaced by the new one. This makes sense in scenarios with randomly generated gradients where a
script is run several times and each time the gradient should be generated again. A basil warning
is thrown instead to let the user know in the console that the old gradient of given name was
replaced by a new one.

The gradient names in question are part of the error/warning messages for easier debugging.

In the array-as-parameter scenario the last argument is saved in var z, since depending on the
number of given arrays it is either the second or third argument.