Readded testing of default colors in test/color-tests.jsx
There is *always* a default color named 'Black' (internal to InDesign, no matter if a local installation shows a different name in the UI), so this test should be run. The reason why the assertion of b.color('Black') failed in the past might be due to the bug described in #73, which has been fixed by now.